Minor refactors
