+++ ire/src/main/scala/hu/bme/mit/ire/listeners/AddListener.scala
@@ -1,3 +1,3 @@
-package hu.bme.mit.ire.nodes.unary
+package hu.bme.mit.ire.listeners
 
 case class AddListener(listener: ChangeListener)
