+++ ire/src/main/scala/hu/bme/mit/ire/nodes/binary/JoinNode.scala
@@ -13,7 +13,7 @@ class JoinNode(override val next: (ReteMessage) => Unit,
 }
 
 class NaturalJoinNode(override val next: (ReteMessage) => Unit, val selector: Vector[Any]
-           ) extends JoinNodeImpl(selector,selector) with SingleForwarder {}
+           ) extends JoinNodeImpl(selector, selector) with SingleForwarder {}
 
 class ParallelJoinNode(override val children: Vector[(ReteMessage) => Unit],
              override val primarySelector: Vector[Any],
@@ -23,16 +23,19 @@ class ParallelJoinNode(override val children: Vector[(ReteMessage) => Unit],
   override def forwardHashFunction(n: TupleType): Int = hashFunction(n)
 
   override def forward(cs: ChangeSet) = super[ForkingForwarder].forward(cs)
+
   override def forward(t: TerminatorMessage) = super[ForkingForwarder].forward(t)
 }
 
 abstract class JoinNodeImpl(val primarySelector: Vector[Any],
               val secondarySelector: Vector[Any])
   extends BinaryNode {
-  val primaryValues = new mutable.HashMap[Vector[Any], mutable.Set[TupleType]]
-  with mutable.MultiMap[Vector[Any], TupleType]
-  val secondaryValues = new mutable.HashMap[Vector[Any], mutable.Set[TupleType]]
-  with mutable.MultiMap[Vector[Any], TupleType]
+  val primaryIndexer =
+  new mutable.HashMap[Vector[Any], mutable.Set[TupleType]]
+    with mutable.MultiMap[Vector[Any], TupleType]
+  val secondaryIndexer =
+  new mutable.HashMap[Vector[Any], mutable.Set[TupleType]]
+    with mutable.MultiMap[Vector[Any], TupleType]
 
   def onPrimary(changeSet: ChangeSet): Unit = {
   val positive = changeSet.positive
@@ -41,28 +44,28 @@ abstract class JoinNodeImpl(val primarySelector: Vector[Any],
   val joinedPositive = for {
     primary <- positive
     key = primarySelector.map(i => primary(i))
-    if secondaryValues.contains(key)
-    secondary <- secondaryValues(key)
-  } yield primary ++ secondary.filterKeys( k => !primary.keySet.contains(k))
+    if secondaryIndexer.contains(key)
+    secondary <- secondaryIndexer(key)
+  } yield primary ++ secondary.filterKeys(k => !primary.keySet.contains(k))
 
   val joinedNegative = for {
     primary <- negative
     key = primarySelector.map(i => primary(i))
-    if secondaryValues.contains(key)
-    secondary <- secondaryValues(key)
-  } yield primary ++ secondary.filterKeys( k => !primary.keySet.contains(k))
+    if secondaryIndexer.contains(key)
+    secondary <- secondaryIndexer(key)
+  } yield primary ++ secondary.filterKeys(k => !primary.keySet.contains(k))
 
   forward(ChangeSet(joinedPositive, joinedNegative))
   positive.foreach(
     m => {
     val key = primarySelector.map(i => m(i))
-    primaryValues.addBinding(key, m)
+    primaryIndexer.addBinding(key, m)
     }
   )
   negative.foreach(
     m => {
     val key = primarySelector.map(i => m(i))
-    primaryValues.removeBinding(key, m)
+    primaryIndexer.removeBinding(key, m)
     }
   )
   }
@@ -74,30 +77,30 @@ abstract class JoinNodeImpl(val primarySelector: Vector[Any],
   val joinedPositive = for {
     secondary <- positive
     key = secondarySelector.map(i => secondary(i))
-    if primaryValues.contains(key)
-    primary <- primaryValues(key)
-  } yield primary ++ secondary.filterKeys( k => !primary.keySet.contains(k))
+    if primaryIndexer.contains(key)
+    primary <- primaryIndexer(key)
+  } yield primary ++ secondary.filterKeys(k => !primary.keySet.contains(k))
 
 
   val joinedNegative = for {
     secondary <- negative
     key = secondarySelector.map(i => secondary(i))
-    if primaryValues.contains(key)
-    primary <- primaryValues(key)
-  } yield primary ++ secondary.filterKeys( k => !primary.keySet.contains(k))
+    if primaryIndexer.contains(key)
+    primary <- primaryIndexer(key)
+  } yield primary ++ secondary.filterKeys(k => !primary.keySet.contains(k))
 
   forward(ChangeSet(joinedPositive, joinedNegative))
 
   positive.foreach(
     m => {
     val key = secondarySelector.map(i => m(i))
-    secondaryValues.addBinding(key, m) //must be used with multimaps
+    secondaryIndexer.addBinding(key, m) //must be used with multimaps
     }
   )
   negative.foreach(
     m => {
     val key = secondarySelector.map(i => m(i))
-    secondaryValues.removeBinding(key, m)
+    secondaryIndexer.removeBinding(key, m)
     }
   )
   }
